FRONT/requirements.txt
PySide6
QT-PyQt-PySide-Custom-Widgets
